% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dental.R
\docType{data}
\name{dental}
\alias{dental}
\title{dental dataset}
\format{
A data frame with 108 observations on the following 4 variables:
  \describe{
    \item{\code{ID}}{patient identifier.}
    \item{\code{Sex}}{sex indicator (0 = female; 1 = male).}
    \item{\code{Time}}{time point raging from 1 to 4.}
    \item{\code{Distance}}{distance (mm) from the center of the pituitary
    gland to the pteryomaxillary fissure.}
  }
}
\description{
The data are from a study of dental growth, presented by Fitzmaurice et al.
(2011). The distance (mm) from the center of the pituitary gland to the
pteryomaxillary fissure were obtained on 11 girls and 16 boys at
ages 8, 10, 12, and 14.
}
\examples{
data(dental)

}
\references{
\itemize{
  \item Fitzmaurice, G. M., Laird, N. M., and Ware, J. H. (2011).
        \emph{Applied Longitudinal Analysis, Second edition}.
        John Wiley and Sons.
  }
}
